+++ NEWS.md
@@ -1,5 +1,10 @@
 # patchwork (development version)
 
+* Axes and axis titles can now be collected using the `plot_layout()` function.
+  Collecting axes will remove duplicated axes in the x- or y-direction. 
+  Collecting axis titles will also remove duplicated titles in the x- or 
+  y-direction, but also merge duplicated titles in the other direction (#150).
+
 # patchwork 1.1.3
 
 * `NULL` can now be used with the different arithmetic operators and will result